Finished frontend ready to implement backend
