+++ Frontend-user-interface/src/App.js
@@ -9,7 +9,7 @@ import Nav from "./pages/Nav";
 import Login from "./pages/Login";
 import Register from "./pages/Register";
 import MyAccount from "./pages/MyAccount";
-import Games from "./pages/Games";
+import GameList from "./pages/GameList";
 import NewGame from "./pages/NewGame";
 
 function App() {
@@ -21,7 +21,7 @@ function App() {
                 <Route path="/login" element={<Login/>}/>
                 <Route path="/register" element={<Register/>}/>
                 <Route path="/myaccount" element={<MyAccount/>}/>
-                <Route path="/games" element={<Games/>}/>
+                <Route path="/games" element={<GameList/>}/>
                 <Route path="/newgame" element={<NewGame/>}/>
             </Route>
         </Routes>
