+++ Frontend-user-interface/src/styles/NewGame.css
@@ -1,8 +1,6 @@
 .new-game-container {
     max-width: 400px;
     margin: auto;
-    padding: 20px;
-    border: 1px solid #ccc;
     border-radius: 5px;
     overflow-y: auto;
 }
